src/requirements.txt
Flask==1.1.2
python-dotenv==0.19.1
python-decouple==3.5